data classの記述場所の変更とレイアウトの編集
