+++ app/src/main/java/com/example/triplan/model/Review.kt
@@ -1,0 +1,6 @@
+package com.example.triplan.model
+
+data class Review(
+    val image: Int,
+    val comment: String
+)
