+++ app/src/main/java/com/example/triplan/model/ReviewGraph.kt
@@ -1,0 +1,6 @@
+package com.example.triplan.model
+
+data class ReviewGraph(
+    val number: Int,
+    val value: Int
+)
