+++ app/src/main/java/com/example/triplan/model/SettingElement.kt
@@ -1,0 +1,6 @@
+package com.example.triplan.model
+
+data class SettingElement(
+    val id: SettingElementId,
+    val name: String
+)
