+++ app/src/main/java/com/example/triplan/model/SettingElementId.kt
@@ -1,0 +1,8 @@
+package com.example.triplan.model
+
+enum class SettingElementId(val value: Int) {
+    EDIT_USER_INFORMATION(0),
+    CONTACT(1),
+    FEEDBACK(2),
+    TERMS_OF_SERVICE(3)
+}
