+++ app/src/main/java/com/example/triplan/model/Traffic.kt
@@ -1,0 +1,6 @@
+package com.example.triplan.model
+
+data class Traffic(
+    val station: String,
+    val detail: String
+)
