+++ app/src/main/java/com/example/triplan/ui/content_plan_traffic/ContentPlanTrafficActivity.kt
@@ -7,12 +7,8 @@ import android.os.Bundle
 import android.os.PersistableBundle
 import androidx.appcompat.app.AppCompatActivity
 import com.example.triplan.R
+import com.example.triplan.model.Traffic
 import kotlinx.android.synthetic.main.activity_content_plan_traffic.*
-
-data class Traffic(
-    val station: String,
-    val detail: String
-)
 
 class ContentPlanTrafficActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
