+++ app/src/main/res/values/dimens.xml
@@ -163,7 +163,7 @@
     <!-- SettingFeedbackActivity -->
     <dimen name="setting_contact_back_margin_start">17dp</dimen>
     <dimen name="setting_contact_back_margin_top">30dp</dimen>
-    <dimen name="setting_contact_title_text_size">30sp</dimen>
+    <dimen name="setting_contact_title_text_size">32sp</dimen>
     <dimen name="setting_contact_title_margin_start">17dp</dimen>
     <dimen name="setting_contact_title_margin_top">17dp</dimen>
     <dimen name="setting_contact_text_height">150dp</dimen>
